Massive documentation updates, and more
Also, PipeAware was just an interface in AS3, but has been implemented
as a class here, so I added some ability for it to handle the startup
process. We'll see how this works, but it could make creating cores much
much easier. It also assumes you're not using a custom Facade subclass.