Updated docs, and renamed PipeAware to PipeAwareCore
Generated docs were not right, so they had to be updated with class
attributes, and some class docs weere wrong or omitted.

IPipeAware was originally an interface in AS3, and not a base class. It
was made a class in this port because there is no interface concept.
The renaming is to make it clear what the actor that is being
subclassed is.