replaced source
Using the plain javascript version as opposed to the puremvc.define()
style. The reason is that the JSDuck documentation compiler does not
understand the PureMVC hierarchy and thus, doesn't show inherited
members. Demos and applications where generated docs are not generated
may freely use the puremvc.define() style, but utilities and framework
classes should always use the native JS prototype style so that the API
docs come out right.